requirements.txt
# Core dependencies for BigQuery + Streamlit application

# Streamlit - Web application framework for data science
streamlit>=1.28.0

# Google Cloud BigQuery - Client library for BigQuery API
google-cloud-bigquery>=3.13.0

# Pandas - Data manipulation and analysis library
pandas>=2.1.0

# DB-dtypes - Required for proper BigQuery data type handling
db-dtypes>=1.1.1

# PyArrow - Improves BigQuery to DataFrame conversion performance
pyarrow>=14.0.0

# python-dotenv - Load environment variables from .env file
python-dotenv>=1.0.0

# LangChain - Framework for building LLM applications
langchain>=0.1.0

# LangChain Google GenAI - Integration for Google's Gemini models
langchain-google-genai>=0.0.6